- Added a dryrun command line option. - Fixed a bug when genre folder didn't exist for file that was encoded   with non-ASCII characters - Tried to fix a bug where all files were not added to iTunes by adding   a 1s sleep

+++ organize_music.py
@@ -1,10 +1,15 @@
 #!/usr/bin/env python
 
-import os, codecs, sys
+import os, codecs, sys, argparse
 from mutagen.id3 import ID3
+from time import sleep
 
-music_path = "/Users/backalor/Music/songs"
-cmus_playlist = "/Users/backalor/Music/playlists/newest-really.m3u"
+music_path = "/Users/brian/Music/songs"
+cmus_playlist = "/Users/brian/Music/playlists/done/newest-really.m3u"
+
+parser = argparse.ArgumentParser(description='Organize music into folders by genre')
+parser.add_argument('--dryrun', nargs='?', default=False, help='Just verify all the proper genre folders exist')
+args = parser.parse_args()
 
 curdir = os.path.abspath(os.getcwd())
 mp3s = [filename for filename in os.listdir(curdir)
@@ -18,29 +23,31 @@ if mp3s:
 			sys.exit()
 		genre = audio['TCON'].text[0]
 		if not os.path.isdir(music_path + "/" + genre):
-			print "ERR: %s directory not found. Fix %s and re-run" % (genre, filename)
+			print "ERR: %s directory not found. Fix %s and re-run" % (genre, filename.decode("utf-8"))
 			sys.exit()
 
-	# Read in the existing playlist from cmus
-	data = codecs.open(cmus_playlist, 'r', "utf-8").read()
+	if args.dryrun==False:
+		# Read in the existing playlist from cmus
+		data = codecs.open(cmus_playlist, 'r', "utf-8").read()
 
-	# Create a new playlist for these new songs
-	playlist = codecs.open("playlist.pl", 'w', "utf-8")
-	for filename in mp3s:
-		audio = ID3(filename)
-		genre = audio['TCON'].text[0]
-		print u"Found Genre: %s for %s" % (genre, filename.decode("utf-8"))
-		if (os.path.isdir(music_path + "/" + genre)):
-			cmd = 'mv "%s" "%s"' % (filename.decode("utf-8"),  music_path + "/" + genre)
-			os.system(cmd.encode("utf-8"))
-			cmd = 'open "%s"' % (music_path + "/" + genre + "/" + filename.decode("utf-8"))
-			os.system(cmd.encode("utf-8"))
-			playlist.write(music_path + "/" + genre + "/" + filename.decode("utf-8"))
-			playlist.write("\n")
-		else:
-			print "ERR: %s directory not found! This should never happen" % genre
-	# Write the existing playlist from cmus at the bottom of the new playlist
-	playlist.write(data)
-	print "Check playlist.pl and if it looks fine manually copy it to ~/.cmus"
+		# Create a new playlist for these new songs
+		playlist = codecs.open("playlist.pl", 'w', "utf-8")
+		for filename in mp3s:
+			audio = ID3(filename)
+			genre = audio['TCON'].text[0]
+			print u"Found Genre: %s for %s" % (genre, filename.decode("utf-8"))
+			if (os.path.isdir(music_path + "/" + genre)):
+				cmd = 'mv "%s" "%s"' % (filename.decode("utf-8"),  music_path + "/" + genre)
+				os.system(cmd.encode("utf-8"))
+				cmd = 'open "%s"' % (music_path + "/" + genre + "/" + filename.decode("utf-8"))
+				os.system(cmd.encode("utf-8"))
+				playlist.write(music_path + "/" + genre + "/" + filename.decode("utf-8"))
+				playlist.write("\n")
+			else:
+				print "ERR: %s directory not found! This should never happen" % genre
+			sleep(1)
+		# Write the existing playlist from cmus at the bottom of the new playlist
+		playlist.write(data)
+		print "Check playlist.pl and if it looks fine manually copy it to ~/.cmus"
 else:
 	print "No .mp3 files found"
